% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chdir.R
\name{chdir}
\alias{chdir}
\title{Characteristic Direction}
\usage{
chdir(ctrl, expm, samples, r = 1)
}
\arguments{
\item{ctrl}{matrix of control data}

\item{expm}{matrix of experiment (perturbed) data}

\item{samples}{vector of drug names}

\item{r}{numeric between 0 and 1. Regularised term, a parameter that smooths
the covariance matrix and reduces potential noise in the dataset. The
default value is 0, no regularisation.}
}
\value{
vector of n-components representing the characteristic direction of
the dataset. \code{n} equals the number of features in the dataset.
b is also a matrix object, and is sorted by its components' absolute
values in descending order.
}
\description{
Calculates the characteristic direction for a phenotypic dataset
}

