% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chdir.R
\name{check_constant_rows}
\alias{check_constant_rows}
\title{error message for constant variance expression data}
\usage{
check_constant_rows(ctrl, expm)
}
\arguments{
\item{ctrl}{matrix of control data}

\item{expm}{matrix of experiment data}
}
\value{
error message if constant variance, otherwise silent
}
\description{
Will stop and return error message if rows have constant variance
}

