% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nipals.R
\name{nipals}
\alias{nipals}
\title{Principal component analysis}
\usage{
nipals(X, a, it = 10, tolerance = 1e-04)
}
\arguments{
\item{X}{matrix}

\item{a}{integer, number of components to calculates}

\item{it}{integer, number of iterations}

\item{tolerance}{float, precision tolerance}
}
\value{
list, T: PCA scores. P: PCA loadings. pcvar: proportion of varation
explained by PCA components
}
\description{
Calculates the principal components accorindg to the algorithm by Wold
}
\examples{
data(iris)
nipals(iris[, 1:4], a = 2)
}

