% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrink_mat.R
\name{shrink_mat}
\alias{shrink_mat}
\title{Shrunken covariance matrix}
\usage{
shrink_mat(R, samples_count, r)
}
\arguments{
\item{R}{PCA scores}

\item{samples_count}{number of samples to explain 95\% of the variance}

\item{r}{regularisation factor to smooth covariance matrix in order to
reduce noise. 1 is no regularisatoin.}
}
\value{
covariance matrix
}
\description{
Calculates a shrunken covariance matrix to be passed as the demoninator
in the LDA formula. The shrunken matrix is in the subspace of the components
that capture 95\% of the variation of the dataset.
}

